# openapi/artifacts/assistant-api.yaml
openapi: 3.0.3
info:
  title: Rapida Assistant API
  version: 0.1.0
paths:
...
  /v1/assistant-deployment/get-all-api-deployment/{assistantId}:
    get:
      operationId: getAllAssistantApiDeployment
      summary: Get all assistant API deployments
      description: Gets API deployments for an assistant. This endpoint requires user authentication with organization and project scope.
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            $ref: "./common.yaml#/components/schemas/Uint64String"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: uint32
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: uint32
            default: 20
        - name: criterias
          in: query
          required: false
          description: JSON encoded array of Criteria objects.
          schema:
            type: string
      responses:
        "200":
          description: Assistant API deployments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllAssistantApiDeploymentResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated request
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "403":
          description: Missing authentication scope
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "500":
          description: Failed to get assistant API deployments
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
  /v1/assistant-deployment/get-all-debugger-deployment/{assistantId}:
    get:
      operationId: getAllAssistantDebuggerDeployment
      summary: Get all assistant debugger deployments
      description: Gets debugger deployments for an assistant. This endpoint requires user authentication with organization and project scope.
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            $ref: "./common.yaml#/components/schemas/Uint64String"
        - name: page
          in: query
          required: false
          schema:
            type: integer
            format: uint32
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            format: uint32
            default: 20
        - name: criterias
          in: query
          required: false
          description: JSON encoded array of Criteria objects.
          schema:
            type: string
      responses:
        "200":
          description: Assistant debugger deployments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllAssistantDebuggerDeploymentResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "401":
          description: Unauthenticated request
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "403":
          description: Missing authentication scope
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
        "500":
          description: Failed to get assistant debugger deployments
          content:
            application/json:
              schema:
                $ref: "./common.yaml#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    GetAllAssistantApiDeploymentResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/AssistantApiDeployment"
        error:
          $ref: "./common.yaml#/components/schemas/Error"
        paginated:
          $ref: "./common.yaml#/components/schemas/Paginated"
    GetAllAssistantDebuggerDeploymentResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/AssistantDebuggerDeployment"
        error:
          $ref: "./common.yaml#/components/schemas/Error"
        paginated:
          $ref: "./common.yaml#/components/schemas/Paginated"
...
    AssistantDebuggerDeployment:
      type: object
      properties:
        id:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        assistantId:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        greeting:
          type: string
        greetingInterruptible:
          type: boolean
        mistake:
          type: string
        inputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        outputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        status:
          type: string
        maxSessionDuration:
          type: integer
          format: uint64
          minimum: 180
          maximum: 600
        idealTimeout:
          type: integer
          format: uint64
          minimum: 15
          maximum: 120
        idealTimeoutBackoff:
          type: integer
          format: uint64
          minimum: 0
          maximum: 5
        idealTimeoutMessage:
          type: string
...
    AssistantApiDeployment:
      type: object
      properties:
        id:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        assistantId:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        greeting:
          type: string
        greetingInterruptible:
          type: boolean
        mistake:
          type: string
        inputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        outputAudio:
          $ref: "#/components/schemas/DeploymentAudioProvider"
        status:
          type: string
        maxSessionDuration:
          type: integer
          format: uint64
          minimum: 180
          maximum: 600
        idealTimeout:
          type: integer
          format: uint64
          minimum: 15
          maximum: 120
        idealTimeoutBackoff:
          type: integer
          format: uint64
          minimum: 0
          maximum: 5
        idealTimeoutMessage:
          type: string
...
    DeploymentAudioProvider:
      type: object
      properties:
        id:
          $ref: "./common.yaml#/components/schemas/Uint64String"
        audioType:
          type: string
        audioProvider:
          type: string
        audioOptions:
          type: array
          items:
            $ref: "./common.yaml#/components/schemas/Metadata"
        status:
          type: string
...